requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
daphne==4.0.0